# services/api/openapi.proposal.yaml
...
openapi: 3.1.0
info:
  title: TuneDeck API
  version: 0.2.0-proposal
servers:
  - url: /
paths:
...
  /v1/me/entitlements:
    get:
      summary: Pro per store and the account's purchases
      description: 'Store-scoped in R1, a purchase on one store unlocks Pro in that store''s app only. Per store the best state wins (verified > pending > revoked > none). Purchases are listed oldest first. `Cache-Control: no-store`.'
      responses:
        '200':
          description: Entitlements
          headers:
            Cache-Control: { schema: { type: string, example: no-store } }
          content: { application/json: { schema: { $ref: '#/components/schemas/BillingEntitlements' } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { description: 'ACCOUNT_DELETING, AUTH_FORBIDDEN or DEVICE_REVOKED', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429':
          description: API_RATE_LIMITED, details.retryAfterSeconds
          headers: { Retry-After: { schema: { type: integer } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '503': { $ref: '#/components/responses/Error' }
  /v1/billing/verify:
    post:
      summary: Send the store's proof after a purchase or restore
      description: |
        Apple: the JWS is verified offline against the pinned Apple root (x5c chain of 3, validity, Apple marker extensions, ES256), then bundle id, Pro product id and environment are checked. Google: the Play Developer API is asked as the service account; a purchased (not pending) purchase is acknowledged after it is stored.
        `appAccountToken` (Apple) or `obfuscatedExternalAccountId` (Google), when set, must be the caller's account id (the `userId` of GET /v1/me). One purchase belongs to one live account; a purchase of a deleted account moves to the account sending it. A revoked purchase stays revoked whatever a later verify says (the answer is then 200 with `state: revoked`). Idempotent: the same proof again returns the stored state.
        Audited `purchase.verified` / `purchase.pending` / `purchase.revoked` when the state or owner changes. Body limit 64 KiB. `Cache-Control: no-store`.
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/BillingVerifyRequest' } } }
      responses:
        '200':
          description: The stored purchase (state verified, pending or revoked)
          headers:
            Cache-Control: { schema: { type: string, example: no-store } }
          content: { application/json: { schema: { $ref: '#/components/schemas/BillingPurchase' } } }
        '400': { description: 'VALIDATION_FAILED (details.field: body, store unsupported, signedTransaction must_be_jws, productId / purchaseToken invalid, unknown_field), or PURCHASE_INVALID with details.reason: signature, wrong_app, unknown_product, environment, malformed, account_unbound, account_mismatch (Apple: appAccountToken must be the signed-in account id); unknown_product, unknown_purchase, cancelled, account_unbound, account_mismatch (Google: obfuscatedExternalAccountId must be the signed-in account id). A purchase bound to an account that was deleted can be restored onto the current one. A purchase the store refunded before it was ever verified stays revoked', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { description: 'ACCOUNT_DELETING, AUTH_FORBIDDEN or DEVICE_REVOKED', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: 'PURCHASE_CONFLICT (the purchase belongs to another live account), IDEMPOTENCY_KEY_REUSED, or IDEMPOTENCY_IN_PROGRESS with Retry-After: 1', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '413': { description: 'PAYLOAD_TOO_LARGE, body over 64 KiB', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429':
          description: API_RATE_LIMITED (30 writes a minute per account), details.retryAfterSeconds
          headers: { Retry-After: { schema: { type: integer } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '503': { description: 'DEPENDENCY_UNAVAILABLE: that store is not configured, or the Google API could not be reached; nothing new is granted', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/webhooks/apple:
    post:
      summary: App Store Server Notifications V2 (called by Apple)
      description: |
        No user token: `signedPayload` (and the `signedTransactionInfo` inside it) is verified like a transaction against the pinned Apple root, and the notification's bundle id must match. Redeliveries are recognized by `notificationUUID` and answered 200 without work. `REFUND` and `REVOKE` revoke the purchase (audited `purchase.revoked` by `system:apple`); `REFUND_REVERSED` whose transaction carries no `revocationDate` makes a revoked purchase verified again (audited `purchase.restored`). Other types and unknown purchases change nothing. A notification whose `data.environment` is not in `APPLE_ENVIRONMENTS` is answered 200 and changes nothing. Handled ids are kept 30 days. Rate limited to 600 requests a minute per client address (429 API_RATE_LIMITED). Body limit 64 KiB.
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/BillingAppleNotification' } } }
      responses:
        '200': { description: 'Accepted (also for redeliveries and notification types that change nothing)', content: { application/json: { schema: { $ref: '#/components/schemas/BillingWebhookAck' } } } }
        '400': { description: 'VALIDATION_FAILED (signedPayload must_be_jws, malformed_body), or PURCHASE_INVALID wrong_app (bundle id mismatch or no notificationUUID)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: 'AUTH_REQUIRED: the signed payload or transaction did not verify, or Apple billing is not configured (answered the same way so a caller cannot tell)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '413': { description: 'PAYLOAD_TOO_LARGE, body over 64 KiB', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/webhooks/google:
    post:
      summary: Play real-time developer notifications via Pub/Sub push (called by Google)
      description: |
        No user token. `Authorization: Bearer <Google OIDC token>` that Pub/Sub attaches: RS256, issuer accounts.google.com, audience `GOOGLE_PUBSUB_PUSH_AUDIENCE`, `email` = `GOOGLE_PUBSUB_PUSH_SERVICE_ACCOUNT` with `email_verified: true`. `message.data` is the base64 notification JSON and its `packageName` must match. Redeliveries are recognized by `message.messageId` (kept 30 days) and answered 204 without work.
        A voided purchase is revoked; a pending purchase that completes (oneTimeProductNotification type 1) is re-checked with the Play API, verified and acknowledged; a pending one cancelled (type 2) is revoked. Anything else (test notifications included) is accepted without change. Rate limited to 600 requests a minute per client address (429 API_RATE_LIMITED). Body limit 64 KiB.
      security: []
      parameters:
        - { name: Authorization, in: header, required: true, description: 'Bearer <Google-signed OIDC token from Pub/Sub push>, not a TuneDeck access token', schema: { type: string } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/BillingPubSubPush' } } }
      responses:
        '204': { description: Accepted (also for redeliveries and notifications that change nothing) }
        '400': { description: 'VALIDATION_FAILED (message malformed, message.data malformed, malformed_body), or PURCHASE_INVALID wrong_app (packageName mismatch)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: 'AUTH_REQUIRED: missing or invalid Pub/Sub token (also when Google''s keys cannot be fetched, or Google billing is not configured)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '413': { description: 'PAYLOAD_TOO_LARGE, body over 64 KiB', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { description: 'DEPENDENCY_UNAVAILABLE: the Play API could not be reached while completing a pending purchase (the notification is retried by redelivery)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
security:
  - bearer: []
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT, description: 'OIDC access token (iss/aud/exp/RS256 verified via JWKS)' }
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 'Optional on signed-in mutations. The same account, method, path and key runs at most once for 24 hours; a repeat with the same body and If-Match replays the first answer with Idempotent-Replayed: true. Different body → 409 IDEMPOTENCY_KEY_REUSED; still running → 409 IDEMPOTENCY_IN_PROGRESS with Retry-After. Ignored where the answer carries a secret (export links, support codes).'
      schema: { type: string, pattern: '^[A-Za-z0-9_.:-]{8,128}$' }
  schemas:
...
    Error:
      type: object
      required: [code, messageKey, requestId, details]
      properties:
        code:
          type: string
          enum: [AUTH_REQUIRED, AUTH_FORBIDDEN, VALIDATION_FAILED, NOT_FOUND, EXPORT_NOT_READY, IDEMPOTENCY_KEY_REQUIRED, PRECONDITION_REQUIRED, REVISION_MISMATCH, PAYLOAD_TOO_LARGE, DEPENDENCY_UNAVAILABLE, REAUTH_REQUIRED, DEVICE_REVOKED, DEVICE_LIMIT, ROLE_REQUIRED, PUBLISH_BLOCKED, CHECK_TOO_SOON, CHECKER_UNAVAILABLE, RIGHTS_ALREADY_REVOKED, API_RATE_LIMITED, ACCOUNT_DELETING, JOB_RETRY_TOO_SOON, JOB_BUSY, JOB_RETRY_LIMIT, JOB_SUPERSEDED, SYNC_RESET_REQUIRED, PURCHASE_INVALID, PURCHASE_CONFLICT, MFA_REQUIRED, IDEMPOTENCY_KEY_REUSED, IDEMPOTENCY_IN_PROGRESS, SUPPORT_ACCESS_REQUIRED, INTERNAL, DIRECTORY_BLOCK_EXISTS]
        messageKey: { type: string, example: errors.settings.revisionMismatch }
        requestId: { type: string }
        details:
          type: object
          description: 'No tokens or PII. VALIDATION_FAILED → {field?, reason, allowed?}; REVISION_MISMATCH → {currentRevision}; REAUTH_REQUIRED → {maxAgeSeconds}; MFA_REQUIRED → {maxAgeSeconds, scope?}; DEVICE_LIMIT → {maxActiveDevices}; PUBLISH_BLOCKED → {reasons}; IDEMPOTENCY_IN_PROGRESS / CHECK_TOO_SOON / JOB_RETRY_TOO_SOON → {retryAfterSeconds}.'
...
    BillingVerifyRequest:
      oneOf:
        - { $ref: '#/components/schemas/BillingVerifyApple' }
        - { $ref: '#/components/schemas/BillingVerifyGoogle' }
      discriminator:
        propertyName: store
        mapping: { apple: '#/components/schemas/BillingVerifyApple', google: '#/components/schemas/BillingVerifyGoogle' }
...
    BillingEntitlements:
      type: object
      required: [pro, purchases]
      properties:
        pro:
          type: object
          required: [apple, google]
          properties:
            apple: { type: string, enum: [verified, pending, revoked, none] }
            google: { type: string, enum: [verified, pending, revoked, none] }
        purchases: { type: array, items: { $ref: '#/components/schemas/BillingPurchase' }, description: Oldest first }
    BillingAppleNotification:
      type: object
      required: [signedPayload]
      properties:
        signedPayload: { type: string, description: 'App Store Server Notifications V2 signed payload (compact JWS, ES256, x5c chain to Apple Root CA - G3)' }
...
    BillingWebhookAck:
      type: object
      required: [ok]
      properties: { ok: { type: boolean, const: true } }
...
  responses:
    Error:
      description: Error envelope
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }